pyproject.toml
[project]
name = "hospitality-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "jinja2>=3.1.6",
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "mcp[cli]>=1.15.0",
    "openai>=1.109.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "uvicorn>=0.37.0",
]
